ongoing attempt to add filewatcher support for .augene files.
It still does not work, as FileWatcher somehow repeatedly invokes
change events (at least on Linux) and Rider debugger fails to catch
the rasher, most likely due to the fact that the runtime cannot
handle anything outside its scope and the SEGV on the Gtk native loop.

It is probably quicker to dump this Gtk# (Xwt)-based solution and find
appropriate language/framework that supports Linux as a first citizen.